change style of deleting notebooks
